requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
httpx==0.28.1
pydantic==2.11.3
qdrant-client==1.13.2
sentence-transformers==3.4.1
PyMuPDF==1.25.3
python-docx==1.1.2
python-multipart==0.0.9
SQLAlchemy==2.0.29
asyncpg==0.29.0
